# api/openapi.yaml
openapi: 3.1.0
info:
  title: Example Server API
  version: 1.0.0
  description: |
    This API allows you to interact with the Example Server.

    **Global Middlewares:**
    - Logging: All requests are logged with request/response details.
...
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - {}
paths:
...
  /orders:
    get:
      summary: Retrieve a list of orders
      description: |
        Retrieves a list of all orders in the system

        **Middlewares applied:**
        - Authentication required
      operationID: listOrders
      tags:
        - Orders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved list of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
                    description: Total number of orders available
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create a new order
      description: |
        Creates a new order in the system

        **Middlewares applied:**
        - Authentication required
      operationID: createOrder
      tags:
        - Orders
      security:
        - bearerAuth: []
      requestBody:
        description: Order details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                  description: Unique identifier for the user placing the order
                products:
                  type: array
                  description: List of products in the order with their quantities
                  items:
                    type: object
                    required:
                      - productId
                      - quantity
                    properties:
                      productId:
                        type: string
                        format: uuid
                        description: Unique identifier for the product
                      quantity:
                        type: integer
                        description: Quantity of the product ordered
                        minimum: 1
      responses:
        '201':
          description: Successfully created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid order data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orders/{orderId}:
    get:
      summary: Get order by ID
      description: |
        Retrieves a single order by its unique identifier

        **Middlewares applied:**
        - Authentication required
      operationId: getOrderById
      tags:
        - Orders
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          description: Unique identifier of the order
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: Invalid order ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update an existing order
      description: |
        Updates details of an existing order.

        **Middlewares applied:**
        - Authentication required
      operationId: updateOrder
      tags:
        - Orders
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          description: Unique identifier of the order to update
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Updated order details
        required: true
        content:
          application/json:
            schema:
              # Should be an array of objects with productId and quantity
              type: object
              properties:
                products:
                  type: array
                  description: List of products in the order with their quantities
                  items:
                    type: object
                    required:
                      - productId
                      - quantity
                    properties:
                      productId:
                        type: string
                        format: uuid
                        description: Unique identifier for the product
                      quantity:
                        type: integer
                        description: Quantity of the product ordered
      responses:
        '200':
          description: Successfully updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid order data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orders/{orderId}/submit:
    post:
      summary: Cancel or Submit an order
      description: |
        Cancels or Submits an existing order
        **Middlewares applied:**
        - Authentication required
      operationId: cancelOrSubmitOrder
      tags:
        - Orders
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          description: Unique identifier of the order to cancel or submit
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Action to perform on the order
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  description: Action to perform on the order
                  enum:
                    - CANCEL
                    - SUBMIT
      responses:
        '200':
          description: Successfully performed action on order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid action or order data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication. Include the token in the Authorization header as "Bearer {token}"
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
          example: "PRODUCT_NOT_FOUND"
        message:
          type: string
          description: Human-readable error message
          example: "The requested product could not be found"
        details:
          type: string
          description: Additional error details
    Order:
      type: object
      required:
        - id
        - productId
        - quantity
        - totalPrice
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the order
        # The products schema should contain an array of objects, the object should be the product ID and the quantity ordered
        products:
          type: array
          description: List of products in the order with their quantities
          items:
            type: object
            required:
              - productId
              - quantity
            properties:
              productId:
                type: string
                format: uuid
                description: Unique identifier for the product
              quantity:
                type: integer
                description: Quantity of the product ordered
                minimum: 1
        totalPrice:
          type: number
          format: float
          description: Total price for the order
          minimum: 0
        accruedLoyaltyPoints:
          type: integer
          description: Loyalty points accrued from this order
          minimum: 0
        orderDate:
          type: string
          format: date-time
          description: Timestamp when the order was placed
        status:
          type: string
          description: Current status of the order
          enum:
            - PENDING
            - PROCESSING
            - SHIPPED
            - DELIVERED
            - CANCELED
...